test/requirements.txt
simplejson
pyaml
requests
qubell-api-python-client

PyMySQL